requirements.txt
django==3.2

django-allauth
python-dateutil

markdown
django-filter
django-resized
python-decouple
django-ckeditor
django-crispy-forms
pillow
six


# python -m pip install --upgrade Pillow >>> install pillow with this